Feature/gfx1250 tr4 load (#9)
* wip: initial support for tr4 ds_load

* Fixed var reference before defining in local read

* Parameterize local read pattern of fp4

* Make local read of fp4 wmma MX-compatible

* Force to use tr load for fp4 if unrollMajor{A, B} == False

* Added NN, NT and TT test cases for fp4 wmma

* Amended comment for tr4-load

+++ projects/hipblaslt/tensilelite/Tensile/AsmMemoryInstruction.py
@@ -44,7 +44,7 @@ class MemoryInstruction:
     numBlocks: int = field(init=False)
     totalWidth: float = field(init=False)
     issueLatency: int = field(init=False)
-    bpe: Optional[int] = field(kw_only=True, default=None)
+    bpe: Optional[float] = field(kw_only=True, default=None)
 
     def __post_init__(self):
         self.numBlocks = 2 if self.numAddresses > 1 or self.numOffsets > 1 else 1
